+++ app/src/main/res/layout/activity_mensajes.xml
@@ -14,28 +14,42 @@
         android:layout_height="wrap_content"
         android:orientation="vertical">
 
-        <FrameLayout
-            android:id="@+id/frameLayout2"
-            android:layout_width="416dp"
-            android:layout_height="578dp"
-            android:layout_marginStart="2dp">
-
-        </FrameLayout>
-
-        <Space
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/containerF2"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
+            android:layout_height="589dp" />
 
         <Space
             android:layout_width="match_parent"
             android:layout_height="47dp" />
 
-        <FrameLayout
-            android:id="@+id/frameLayout1"
+        <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="68dp">
+            android:layout_height="match_parent"
+            android:orientation="horizontal">
 
-        </FrameLayout>
+            <EditText
+                android:id="@+id/mensajeLayout"
+                android:layout_width="366dp"
+                android:layout_height="wrap_content"
+                android:layout_gravity="left"
+                android:hint="Escribe un Mensaje..."
+                android:nextFocusLeft="@id/enviarBtn">
+
+            </EditText>
+
+            <ImageButton
+                android:id="@+id/enviarBtn"
+                android:layout_width="40dp"
+                android:layout_height="40dp"
+                android:layout_gravity="right"
+                android:layout_weight="1"
+                android:background="@drawable/ic_action_name"
+                android:nextFocusRight="@id/mensajeLayout"
+                android:onClick="enviarMensaje"
+                android:text="Enviar" />
+        </LinearLayout>
 
         <Button
             android:id="@+id/button6"
