--- app/src/main/res/layout/fragment_one.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".OneFragment">
-
-    <EditText
-        android:id="@+id/cuadradoMensaje"
-        android:layout_width="366dp"
-        android:layout_height="wrap_content"
-        android:hint="Escribe un Mensaje..."
-        android:nextFocusLeft="@id/enviarBtn">
-
-    </EditText>
-
-    <ImageButton
-        android:id="@+id/enviarBtn"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="right"
-        android:layout_weight="1"
-        android:background="@drawable/ic_action_name"
-        android:nextFocusRight="@id/cuadradoMensaje"
-        android:text="Enviar" />
-
-
-</FrameLayout>
-
